Potential fix for code scanning alert no. 35: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to avoid overflow. We can achieve this by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done using the larger type, and the result will not overflow.

Specifically, we will cast buf->stride to size_t before multiplying it by iter.height on line 354.

+++ libwebp-1.0.3/src/demux/anim_decode.c
@@ -351,7 +351,7 @@ int WebPAnimDecoderGetNext(WebPAnimDecoder* dec,
     WebPDecoderConfig* const config = &dec->config_;
     WebPRGBABuffer* const buf = &config->output.u.RGBA;
     buf->stride = NUM_CHANNELS * width;
-    buf->size = buf->stride * iter.height;
+    buf->size = (size_t)buf->stride * iter.height;
     buf->rgba = dec->curr_frame_ + out_offset;
 
     if (WebPDecode(in, in_size, config) != VP8_STATUS_OK) {
